Move filterState in its own directory and extract reducer
